Last minute changes, changed LLM to 16bit

+++ attention_lens/model/get_model.py
@@ -21,7 +21,7 @@ def get_model(
         The light-weight hooked model and tokenizer.
     """
 
-    model = AutoModelForCausalLM.from_pretrained(model_name)
+    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16)
     # model.to(device)
 
     model.eval()
